# demo-app/docs/API_SPEC.yaml
openapi: "3.1.0"
info:
  title: Demo App API
  version: 1.0.0
  description: |
    REST API for the Angular 21 Enterprise Demo App.

    ## Authentication
    All protected endpoints use an HttpOnly JWT cookie (`access-token`, RS256, 15-min TTL)
    set by `POST /auth/login` and rotated by `POST /auth/refresh`.
    The browser sends it automatically when Angular `HttpClient` is configured with
    `withCredentials: true`. No `Authorization` header is used.

    ## CSRF
    CSRF tokens are **disabled**. Protection is provided by `SameSite=Strict` on both
    auth cookies combined with the strict CORS origin allowlist.

    ## Pagination
    List endpoints accept `page` (0-based) and `pageSize` query parameters.
    The total item count is returned in the `X-Total-Count` response header.
    Response bodies for list endpoints are plain JSON arrays.

    ## Errors
    All error responses share the same shape:
    ```json
    { "code": "ERROR_CODE", "message": "human-readable description",
      "fields": [{ "field": "email", "message": "must be a valid email" }] }
    ```
servers:
  - url: https://api.company.com/api/v1
    description: Production
  - url: https://staging.company.com/api/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - cookieAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with email and password
      operationId: login
      security: []
      description: |
        Validates credentials and sets two HttpOnly cookies:
        - `access-token` (JWT RS256, 15 min, SameSite=Strict)
        - `refresh-token` (opaque 256-bit, 7 days, SameSite=Strict, Path=/auth/refresh)

        A non-HttpOnly `XSRF-TOKEN` cookie is **not** issued (CSRF is disabled).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful. Cookies set in `Set-Cookie` headers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Log out and revoke the refresh token
      operationId: logout
      description: Clears `access-token` and `refresh-token` cookies and revokes the refresh token in the database.
      responses:
        '204':
          description: Logged out successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Silently rotate auth cookies
      operationId: refreshToken
      security: []
      description: |
        Uses the HttpOnly `refresh-token` cookie (scoped to `Path=/auth/refresh`).
        Rotates both cookies and revokes the old refresh token.
        Called automatically by the Angular HTTP interceptor on receiving a 401.
      responses:
        '200':
          description: Cookies rotated. New cookies set in `Set-Cookie` headers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/password:
    post:
      tags: [Auth]
      summary: Change the authenticated user's password
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: Validation error or wrong current password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                wrongPassword:
                  value:
                    code: WRONG_PASSWORD
                    message: Current password is incorrect
                tooShort:
                  value:
                    code: VALIDATION_ERROR
                    message: Validation failed
                    fields:
                      - field: newPassword
                        message: must be at least 6 characters

  # ── Users ─────────────────────────────────────────────────────────────────
...
  /roles:
    get:
      tags: [Roles]
      summary: List roles
      operationId: listRoles
      x-permission: rolesView
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated role list
          headers:
            X-Total-Count:
              $ref: '#/components/headers/XTotalCount'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Roles]
      summary: Create a role
      operationId: createRole
      x-permission: rolesEdit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleRequest'
      responses:
        '201':
          description: Role created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Role name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: CONFLICT
                message: A role with this name already exists
        '422':
          $ref: '#/components/responses/ValidationError'
  /roles/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'

    patch:
      tags: [Roles]
      summary: Update a role
      operationId: updateRole
      x-permission: rolesEdit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleRequest'
      responses:
        '200':
          description: Updated role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Cannot modify a built-in role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: BUILT_IN_ROLE
                message: Built-in roles cannot be modified
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Roles]
      summary: Delete a role
      operationId: deleteRole
      x-permission: rolesEdit
      responses:
        '204':
          description: Role deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Built-in role or role still assigned to users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                builtIn:
                  value:
                    code: BUILT_IN_ROLE
                    message: Built-in roles cannot be deleted
                hasUsers:
                  value:
                    code: CONFLICT
                    message: Cannot delete a role that is assigned to users

  # ── Categories ────────────────────────────────────────────────────────────
  /categories:
    get:
      tags: [Categories]
      summary: List document categories
      operationId: listCategories
      description: Returns only categories where the caller's role has `canView = true`.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated category list
          headers:
            X-Total-Count:
              $ref: '#/components/headers/XTotalCount'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentCategory'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Categories]
      summary: Create a category
      operationId: createCategory
      x-permission: rolesEdit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentCategory'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Slug already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: CONFLICT
                message: A category with this slug already exists
        '422':
          $ref: '#/components/responses/ValidationError'
  /categories/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'

    patch:
      tags: [Categories]
      summary: Update a category
      operationId: updateCategory
      x-permission: rolesEdit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '200':
          description: Updated category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentCategory'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Categories]
      summary: Delete a category
      operationId: deleteCategory
      x-permission: rolesEdit
      responses:
        '204':
          description: Category deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Category still contains documents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: CONFLICT
                message: Cannot delete a category that contains documents
...
  /documents:
    get:
      tags: [Documents]
      summary: List documents in a category
      operationId: listDocuments
      description: Returns ready documents the caller's role has `canView = true` for in the given category.
      parameters:
        - name: categoryId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Filter by category
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Paginated document list
          headers:
            X-Total-Count:
              $ref: '#/components/headers/XTotalCount'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller's role has `canView = false` for this category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /documents/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'

    delete:
      tags: [Documents]
      summary: Delete a document
      operationId: deleteDocument
      description: Marks the document as deleted and schedules the S3 object for removal.
      responses:
        '204':
          description: Document deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller's role has `canDelete = false` for this document's category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Reports ───────────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access-token
      description: |
        HttpOnly JWT cookie (RS256, 15-min TTL). Set by `POST /auth/login` and
        rotated by `POST /auth/refresh`. Angular `HttpClient` sends it automatically
        when configured with `withCredentials: true`. Never accessible to JavaScript.

  # ── Headers ───────────────────────────────────────────────────────────────
  headers:
    XTotalCount:
      description: Total number of items matching the query (before pagination)
      schema:
        type: integer
        example: 42

  # ── Parameters ────────────────────────────────────────────────────────────
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Resource UUID
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Zero-based page number
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        enum: [5, 10, 25, 50]
        default: 10
      description: Number of items per page
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication cookie
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: UNAUTHORIZED
            message: Authentication required
    Forbidden:
      description: Authenticated but insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: FORBIDDEN
            message: You do not have permission to perform this action
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: NOT_FOUND
            message: The requested resource was not found
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: VALIDATION_ERROR
            message: Validation failed
            fields:
              - field: email
                message: must be a valid email address
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
            example: 60
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: RATE_LIMITED
            message: Too many requests. Please wait before retrying.
...
  schemas:
...
    UserStatus:
      type: string
      enum: [active, inactive, suspended]
      description: Lifecycle status of a user account
    DocumentStatus:
      type: string
      enum: [pending, ready, deleted]
      description: |
        - `pending`: upload presigned URL issued, S3 object not yet confirmed
        - `ready`: upload committed, document accessible
        - `deleted`: soft-deleted, S3 object scheduled for removal
      readOnly: true
    Theme:
      type: string
      enum: [light, dark, system]
      description: UI colour theme preference. `system` resolves via `prefers-color-scheme`.
...
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - BUILT_IN_ROLE
        - WRONG_PASSWORD
        - UPLOAD_EXPIRED
        - RATE_LIMITED
        - INTERNAL_ERROR

    # ── Core Entities ────────────────────────────────────────────────────────
    RolePermissions:
      type: object
      required:
        - usersView
        - usersCreate
        - usersEdit
        - usersDelete
        - rolesView
        - rolesEdit
      properties:
        usersView:
          type: boolean
        usersCreate:
          type: boolean
        usersEdit:
          type: boolean
        usersDelete:
          type: boolean
        rolesView:
          type: boolean
        rolesEdit:
          type: boolean
    User:
      type: object
      required: [id, email, displayName, status, roleId, roleName, permissions, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        email:
          type: string
          format: email
        displayName:
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'
        roleId:
          type: string
          format: uuid
        roleName:
          type: string
        permissions:
          $ref: '#/components/schemas/RolePermissions'
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    Role:
      type: object
      required: [id, name, description, isBuiltIn, permissions, userCount, createdAt]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        description:
          type: string
        isBuiltIn:
          type: boolean
          readOnly: true
          description: Built-in roles (Administrator, Manager, Viewer) cannot be modified or deleted
        permissions:
          $ref: '#/components/schemas/RolePermissions'
        userCount:
          type: integer
          readOnly: true
          description: Number of users currently assigned to this role
        createdAt:
          type: string
          format: date-time
          readOnly: true
    RoleRequest:
      type: object
      required: [name, permissions]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
        description:
          type: string
          maxLength: 500
        permissions:
          $ref: '#/components/schemas/RolePermissions'
    CategoryVisibilityItem:
      type: object
      required: [roleId, roleName, canView, canUpload, canDelete]
      properties:
        roleId:
          type: string
          format: uuid
        roleName:
          type: string
        canView:
          type: boolean
        canUpload:
          type: boolean
        canDelete:
          type: boolean
    DocumentCategory:
      type: object
      required: [id, name, slug, sortOrder, documentCount, visibility, createdAt]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        slug:
          type: string
        sortOrder:
          type: integer
        documentCount:
          type: integer
          readOnly: true
        visibility:
          type: array
          items:
            $ref: '#/components/schemas/CategoryVisibilityItem'
        createdAt:
          type: string
          format: date-time
          readOnly: true
    CategoryRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
          description: URL-safe identifier. Auto-generated from name if omitted.
        sortOrder:
          type: integer
          minimum: 0
          default: 0
...
    Document:
      type: object
      required: [id, categoryId, categoryName, ownerId, ownerName, filename, mimeType, sizeBytes, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        categoryId:
          type: string
          format: uuid
        categoryName:
          type: string
        ownerId:
          type: string
          format: uuid
        ownerName:
          type: string
        filename:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/DocumentStatus'
        createdAt:
          type: string
          format: date-time
          readOnly: true
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
          format: password
    LoginResponse:
      type: object
      required: [user, permissions, theme]
      properties:
        user:
          $ref: '#/components/schemas/User'
        permissions:
          $ref: '#/components/schemas/RolePermissions'
        theme:
          $ref: '#/components/schemas/Theme'
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          format: password
        newPassword:
          type: string
          minLength: 6
          format: password
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Human-readable explanation
        fields:
          type: array
          description: Per-field validation errors (present only for VALIDATION_ERROR)
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
                description: JSON path to the invalid field
              message:
                type: string
                description: Why the field failed validation

    # ── Health ───────────────────────────────────────────────────────────────
...